# Packer_initialize: lazily allocate the extension and cache hashes