Each packer eagerly allocate two hashes, one for it's extension registry another to act as a cache for lookup and it's expensive.

Assuming no extension type is used, we can significantly speedup the packer instantiation by skipping these two allocations.

And in case extension types are used, we can still delay the cache allocation until it is actually needed.

benchmark (`Object.new` is used as a comparison point):

```ruby
require "benchmark/ips"
require "msgpack"

FACTORY = MessagePack::Factory.new
Benchmark.ips do |x|
  x.report("Object.new") { Object.new }
  x.report("FACTORY.packer") { FACTORY.packer }
  x.compare!
end
```

before:
```
Calculating -------------------------------------
          Object.new      8.512M (± 0.7%) i/s -     42.709M in   5.017901s
      FACTORY.packer      1.932M (± 0.8%) i/s -      9.697M in   5.019840s

Comparison:
          Object.new:  8511619.9 i/s
      FACTORY.packer:  1931874.0 i/s - 4.41x  (± 0.00) slower
```

after:
```
Calculating -------------------------------------
          Object.new      8.484M (± 1.7%) i/s -     42.680M in   5.032470s
      FACTORY.packer      3.399M (± 1.7%) i/s -     17.159M in   5.049362s

Comparison:
          Object.new:  8483734.5 i/s
      FACTORY.packer:  3399232.3 i/s - 2.50x  (± 0.00) slower
```